completed JobType for jenkinsService.getJobs
